LPS-34882 BatchSessionImpl's delete logic is different from generated *PersistenceImpl.java

+++ portal-impl/src/com/liferay/portal/service/persistence/BatchSessionImpl.java
@@ -34,7 +34,9 @@ public class BatchSessionImpl implements BatchSession {
 				model.getClass(), model.getPrimaryKeyObj());
 		}
 
-		session.delete(model);
+		if (model != null) {
+			session.delete(model);
+		}
 	}
 
 	public boolean isEnabled() {
